#33_feat_impl final animation, layout by design guide
