+++ lib/onboard/animate/picker/picker_data.dart
@@ -1,0 +1,22 @@
+const PickerData = '''
+[
+    [
+        "오전",
+        "오후"
+    ],
+    [
+        1,
+        2,
+        3,
+        4,
+        5,
+        6,
+        7,
+        8,
+        9,
+        10,
+        11,
+        12
+    ]
+]
+    ''';
